Avoid cargo always re-compiling crates that use the syntax extension
The filename passed to new_parser_from_source_str gets added to the
codemap, which is used to generate the dep-info list, so cargo thought the
crate depended on a non-existent file called "peg". The dep-info output
ignores filenames surrounded by "<>".

It should probably instead somehow use the crate map entry for the
.rustpeg file, but that can wait for the new plugin API, if it applies
there.