Optional arguments for rule expressions
This allows a rule to override global value of a grammar argument.
Since grammar arguments are propagated recursively through rules, change
would affect only a single 'call stack'. This would help to implement
context sensitive grammars without duplicating rules.

One example where this might be useful, is parsing Rust itself -
position of the expression in the source text determines if struct
literals are allowed: an `if` condition can be any expression except for
a struct literal.