Remove error checks for mixing up PEG rule and Rust value arguments
When a rule with PEG rule closure arguments is recursive, it can be
useful to pass down the rule argument as a value, since re-capturing it
results in a type-length limit error from rustc. (#226)

Rustc's type error if you do pass the wrong type isn't as good as the
custom one, but isn't too bad.