Require ParseElem::Element: Copy
Tokens/characters should be cheap to copy because they are copied often.
The previous commit enables ParseElem implementations to return a
reference to an element, and references are `Copy`.

This is a breaking change for implementers of custom `ParseElem` types
or users of the `ParseElem` impl for non-Copy `[T]`, but the vast
majority of grammars use `for str`, which is not affected.