Add error recovery
Adds a new expression `error!{"message" e}` which reports an error and
attempts to recover it with `e`. Also `error_if!{}` and
`error_unless!{}`.

Maintains and returns a list of errors which have been recovered.

This enables the parser to report multiple errors, which is useful
in contexts like IDEs and compilers where the user may find it easier
to be presented with all errors rather than just the first.

This commit contains a manually-updated grammar.rs,
sufficient only for bootstrapping but with no support for
parsing the new expression types.
The next commit contains the bootstrapped one.