Allow combining `#[cache]` with arguments
This is a naïve first attempt at enabling caching with arguments,
supporting both values and ownable references. References are less
efficient because it is extremely difficult to satisfy the
constraint `K: Borrow<Q>` for `HashMap` with a tuple. Most
recommendations seem to be to either do some contortions with trait
objects, or to use the hashbrown crate and implement
`hashbrown::Equivalent`, but I have not been able to successfully
think through how to do this in a way which works with arbitrary
tuples of foreign types. Still, a half-measure which can be
improved later is hopefully better than nothing.